requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
cssutils==2.9.0
decouple==0.0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-admin-interface==0.26.0
django-colorfield==0.10.0
django-import-export==3.2.0
emoji==2.10.1
et-xmlfile==1.1.0
future==1.0.0
MarkupPy==1.14
mysqlclient==2.2.4
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
pynliner==0.8.0
python-decouple==3.8
python-slugify==8.0.1
PyYAML==6.0.1
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
